pyproject.toml
[project]
name = "lichess-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "backoff~=2.2",
    "chess~=1.11",
    "pyyaml~=6.0",
    "requests~=2.32",
    "rich~=14.0",
    "openai",
    "anthropic",
    "python-dotenv",
    "flask>=3.1.1",
]

[tool.setuptools.packages]
find = { where = ["lib"] }
